requirements.txt
djangorestframework
markdown       
django-filter  
psycopg2
xmltodict
requests